naming-jsx-engine/requirements.txt
tqdm
PyGithub
requests
bs4
python-dotenv